Decouple link validation into axios service
